+++ main/main/Functions.h
@@ -2,6 +2,6 @@
 #define FUNCTIONS_H
 #include <vector>
 
-float dotProduct(std::vector<int> weight, std::vector<int> input);
+float dotProduct(std::vector<int> weight, std::vector<float> input);
 
 #endif
